games/requirements.txt
asgiref==3.9.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cloudinary==1.44.1
cryptography==45.0.5
Django==5.2.4
django-ckeditor==6.7.3
django-cloudinary-storage==0.3.0
django-js-asset==3.1.2
django-oauth-toolkit==3.0.1
djangorestframework==3.16.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
jwcrypto==1.5.6
mysqlclient==2.2.7
oauthlib==3.3.1
packaging==25.0
pycparser==2.22
PyMySQL==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
